Make some progress on comments
