+++ packages/backend/src/controllers/tasks/dto/task-response-dto.ts
@@ -1,0 +1,4 @@
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { TaskModel } from '../../../db/task/task.schema';
+
+export class TaskResponseDto extends TaskModel {}
