+++ packages/backend/src/controllers/tasks/dto/update-house-tasks.dto.ts
@@ -3,8 +3,10 @@ import { ApiPropertyOptional } from '@nestjs/swagger';
 export default class UpdateHouseTasksDto {
   @ApiPropertyOptional()
   name: string;
+
   @ApiPropertyOptional()
   description: string;
+
   @ApiPropertyOptional()
   pool: string[];
 }
